@@ -1,4 +1,4 @@
-package xyz.luan.example
+package xyz.luan.audioplayers_example
 
 import androidx.annotation.NonNull;
 import io.flutter.embedding.android.FlutterActivity
